pyproject.toml
[tool.poetry]
name = "shopware-knowledge"
version = "0.1.0"
description = ""
authors = ["Dominic Klein <d.klein@shopware.com>"]
readme = "README.md"
packages = [{include = "web"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
langchain = "^0.0.91"
faiss-cpu = "^1.7.3"
unstructured = "^0.4.11"
libmagic = "^1.0"
openai = "^0.26.5"
tiktoken = "^0.2.0"
fastapi = "^0.92.0"
pydantic = "^1.10.5"
uvicorn = "^0.20.0"
gunicorn = "^20.1.0"
pytest = "^7.2.1"
httpcore = {extras = ["http2"], version = "^0.16.3"}
httpx = "^0.23.3"
tensorflow-text = "^2.11.0"
python-multipart = "^0.0.5"
aiofiles = "^23.1.0"
beautifulsoup4 = "^4.11.2"
lxml = "^4.9.2"
python-frontmatter = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
